+++ src/main/java/org/jacobarchambault/travelexpenses/amounts/Amount.java
@@ -1,5 +1,5 @@
 package org.jacobarchambault.travelexpenses.amounts;
 
 public interface Amount {
-	int total();
+	double total();
 }
